requirements.txt
# Data Sources
entsoe-py==0.6.8
requests==2.31.0

# ML & Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Feature Store & Model Registry
hopsworks==3.7.0

# Monitoring & Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
tenacity==8.2.3

# Web UI
streamlit==1.29.0
joblib==1.3.2

# Testing
pytest==7.4.3

